requirements.txt
cdktf~=0.8.1
cdktf-cdktf-provider-aws
python-dotenv
